+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * jdbc: @NoTransaction can now be used on jobs
 * jdbc: fixed usage of multiple different DataSources when there is an active transaction
 * json: TSGenerator will now use more type-safe string template types for java.time classes, e.g. `${number}-${number}-${number}` instead of `string`
-* openapi: fix compliancy: output parameter types as lowercase in OpenAPI spec
+* jackson: serialize enums using their toString() method by default, this fixes `openapi` module usage with `jackson`
 * server: `Server(InetSocketAddress(0))` can now be used to bind to any available port.
 
 # 1.6.9